Move MySQLVisitor and PostgreSQLVisitor to visitor package
